cq: Move internal modules to support directory
Fixes #474.
