+++ pyatv/support/cache.py
