+++ tests/airplay/test_airplay_auth.py
@@ -8,7 +8,7 @@ from aiohttp.test_utils import AioHTTPTestCase, unittest_run_loop
 from pyatv.airplay import srp
 from pyatv.airplay.auth import DeviceAuthenticator, AuthenticationVerifier
 from pyatv.exceptions import AuthenticationError
-from pyatv.net import HttpSession
+from pyatv.support.net import HttpSession
 from tests.airplay.fake_airplay_device import (
     FakeAirPlayDevice,
     DEVICE_IDENTIFIER,
